Fix provider URL, button types, pool teardown; document the login-CSRF gap
Security gap I introduced by omission, now documented rather than half-fixed:

- [CREDIT:@Copilot] The CSRF gate keys off the HTTP method and treats GET as
  safe. That assumption does not hold here: /api/auth/magic-link/verify is a
  public GET that calls createSession and sets the session cookie, so an
  attacker can request a magic link for their own account, send the victim that
  URL, and have the victim land signed into the attacker's account — anything
  they do next belongs to the attacker. /api/auth/verify-email is a public
  state-changing GET too, at lower severity.

  Method-based CSRF protection cannot close this. The standard fix is an
  interstitial that turns the emailed link into a same-origin POST, which
  changes what the link does and is therefore a product decision, not a patch.
  Recorded at the gate in middleware.ts and in CLAUDE.md's Security Invariants,
  because that section previously read as though the whole app were covered.

Fixes:
- [CREDIT:@Copilot] core/ai/completion.ts pointed Google at
  generativelanguage.googleapis.com/v1beta, producing /v1beta/chat/completions.
  The OpenAI-compatible surface is under /v1beta/openai, so every Gemini request
  404'd — while providers.test.ts asserted the mapping as if it worked.
- [CREDIT:@Copilot] maxTokens accepted fractional values that pass validation
  and then fail at the provider; now z.number().int().
- [CREDIT:@Copilot] CopyButton, Modal and Dropdown render bare <button>, which
  defaults to type="submit" and submits any enclosing form. All now explicit.
- [CREDIT:@Copilot] closeTestDb only closed the harness pool, leaving the pool
  the code under test opened alive with its keepalive timers. Added closeDb() to
  core/db/client and called it during integration teardown.
- [CREDIT:@Copilot] contentSecurityPolicy had no test despite headers.test.ts
  existing. Added cases for the protective directives, the Stripe origins, and
  that 'unsafe-eval' appears only when dev is true.

+++ app/api/ai/completion/route.ts
@@ -22,7 +22,7 @@ const completionSchema = z.object({
     .min(1)
     .max(MAX_MESSAGES),
   temperature: z.number().min(0).max(2).optional(),
-  maxTokens: z.number().min(1).optional(),
+  maxTokens: z.number().int().min(1).optional(),
 })
 
 export const POST = withErrorHandler(async (request: Request) => {
